+++ pyairmore/__init__.py
@@ -2,7 +2,7 @@
 Pyairmore package
 """  # todo [1] package doc
 
-__version__ = "0.1.0a2"
+__version__ = "0.1.0a3"
 __author__ = "Eray Erdin"