+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.0a3] - 2018-10-16
 ### Changed
  - `DeviceService::fetch_device_detail` returns `DeviceDetail` instance
  - `Service::request` now returns `requests.Response`